+++ testing/T1QuerySerialization.hs
@@ -31,7 +31,8 @@ import ClickHaskell
 -- GHC included
 import Control.Monad (when)
 import Data.ByteString as BS (singleton)
-import Data.ByteString.Builder (toLazyByteString)
+import Data.ByteString.Char8 as BS8 (pack)
+import Data.ByteString.Builder (toLazyByteString, byteString)
 import GHC.Generics (Generic)
 
 
@@ -47,9 +48,9 @@ t1 conn = do
   runTestForType @UInt64 conn [minBound, toEnum 0, maxBound]
   runTestForType @UUID conn [minBound, toEnum 0, maxBound]
   runTestForType @ChString conn (map (toChType . BS.singleton) [1..255])
-  -- ToDo: querySerializationTest @(LowCardinality ChString) connection (map (toChType . BS.singleton) [0..255])
-  -- ToDo: querySerializationTest @(ChArray ChString) connection [toChType $ map BS.singleton [0..255]]
-  -- ToDo: querySerializationTest @(ChArray ChInt64) connection [toChType [0 :: ChInt64 .. 255]]
+  -- ToDo: runTestForType @(LowCardinality ChString) connection (map (toChType . BS.singleton) [0..255])
+  -- ToDo: runTestForType @(ChArray ChString) connection [toChType $ map BS.singleton [0..255]]
+  -- ToDo: runTestForType @(ChArray ChInt64) connection [toChType [0 :: ChInt64 .. 255]]
 
 
 runTestForType ::
@@ -64,7 +65,7 @@ runTestForType ::
   =>
   Connection -> [chType] -> IO ()
 runTestForType connection testValues = do
-  let typeName = chTypeName @chType
+  let typeName = (byteString . BS8.pack) (chTypeName @chType)
   mapM_
     (\chType -> do
       [selectChType] <-